fix: correção no db
