--- src/back-end/alembic/versions/b51db86cb01e_database_v1.py
@@ -1,120 +1,0 @@
-"""database v1
-
-Revision ID: b51db86cb01e
-Revises: 81b3675628a6
-Create Date: 2025-05-27 16:45:26.688505
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b51db86cb01e'
-down_revision: Union[str, None] = '81b3675628a6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    
-    op.drop_table('model')
-    op.drop_table('building')
-    op.execute('DROP TABLE "user" CASCADE')
-    op.execute('DROP TABLE "result" CASCADE')
-    op.execute('DROP TABLE "expedition" CASCADE')
-    op.execute('DROP TABLE "audit" CASCADE')
-    op.execute('DROP TABLE "image" CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('fissure',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('confiabilidade', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('categoria', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('id_image', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['id_image'], ['image.id'], name=op.f('fissure_id_image_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('fissure_pkey')),
-    sa.UniqueConstraint('confiabilidade', name=op.f('fissure_confiabilidade_key'), postgresql_include=[], postgresql_nulls_not_distinct=False)
-    )
-    op.create_table('image',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('image_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('url', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('hora_coleta', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('orientacao', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('id_predio', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('img_resultado', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('anotacao', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('confiabilidade', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('id_modelo', sa.INTEGER(), autoincrement=False, nullable=True),
-    sa.ForeignKeyConstraint(['id_predio'], ['expedition.id'], name='image_id_predio_fkey'),
-    sa.PrimaryKeyConstraint('id', name='image_pkey'),
-    postgresql_ignore_search_path=False
-    )
-    op.create_table('audit',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('id_image', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('id_user', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('id_resultado', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.ForeignKeyConstraint(['id_image'], ['image.id'], name=op.f('audit_id_image_fkey')),
-    sa.ForeignKeyConstraint(['id_resultado'], ['result.id'], name=op.f('audit_id_resultado_fkey')),
-    sa.ForeignKeyConstraint(['id_user'], ['user.id'], name=op.f('audit_id_user_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('audit_pkey')),
-    sa.UniqueConstraint('id_resultado', name=op.f('audit_id_resultado_key'), postgresql_include=[], postgresql_nulls_not_distinct=False)
-    )
-    op.create_table('expedition',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('expedition_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('localizacao', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('data_criacao', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('ultima_att', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('id_responsavel', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('descricao', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.Column('foto_capa', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.ForeignKeyConstraint(['id_responsavel'], ['user.id'], name='expedition_id_responsavel_fkey'),
-    sa.PrimaryKeyConstraint('id', name='expedition_pkey'),
-    postgresql_ignore_search_path=False
-    )
-    op.create_table('result',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('resultado', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('result_pkey'))
-    )
-    op.create_table('user',
-    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('user_id_seq'::regclass)"), autoincrement=True, nullable=False),
-    sa.Column('email', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('nome_completo', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('senha', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('cargo', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name='user_pkey'),
-    sa.UniqueConstraint('email', name='user_email_key', postgresql_include=[], postgresql_nulls_not_distinct=False),
-    postgresql_ignore_search_path=False
-    )
-    op.create_table('building',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('id_expedicao', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('complemento', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('descricao', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.Column('foto_fachada', sa.VARCHAR(), autoincrement=False, nullable=True),
-    sa.ForeignKeyConstraint(['id_expedicao'], ['expedition.id'], name=op.f('building_id_expedicao_fkey')),
-    sa.PrimaryKeyConstraint('id', name=op.f('building_pkey'))
-    )
-    op.create_table('model',
-    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
-    sa.Column('url', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
-    sa.Column('tipo', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.Column('loss', sa.INTEGER(), autoincrement=False, nullable=False),
-    sa.Column('loss_tipo', sa.VARCHAR(), autoincrement=False, nullable=False),
-    sa.PrimaryKeyConstraint('id', name=op.f('model_pkey'))
-    )
-    # ### end Alembic commands ###
