+++ src/back-end/alembic/versions/d0e630350e8a_criando_tabelas_iniciais.py
@@ -1,0 +1,100 @@
+"""Criando tabelas iniciais
+
+Revision ID: d0e630350e8a
+Revises: 
+Create Date: 2025-05-28 17:33:23.329192
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = 'd0e630350e8a'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('fissure')
+    op.drop_table('image')
+    op.drop_table('building')
+    op.drop_table('expedition')
+    op.drop_table('user')
+    op.drop_table('model')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('fissure',
+    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
+    sa.Column('confiabilidade', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.Column('categoria', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('id_image', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.ForeignKeyConstraint(['id_image'], ['image.id'], name=op.f('fissure_id_image_fkey')),
+    sa.PrimaryKeyConstraint('id', name=op.f('fissure_pkey'))
+    )
+    op.create_table('model',
+    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('model_id_seq'::regclass)"), autoincrement=True, nullable=False),
+    sa.Column('url', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('tipo', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('loss', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.Column('loss_tipo', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.PrimaryKeyConstraint('id', name='model_pkey'),
+    postgresql_ignore_search_path=False
+    )
+    op.create_table('image',
+    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
+    sa.Column('url', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('hora_coleta', postgresql.TIMESTAMP(), autoincrement=False, nullable=False),
+    sa.Column('orientacao', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('id_predio', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.Column('img_resultado', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('anotacao', sa.VARCHAR(), autoincrement=False, nullable=True),
+    sa.Column('id_modelo', sa.INTEGER(), autoincrement=False, nullable=True),
+    sa.ForeignKeyConstraint(['id_modelo'], ['model.id'], name=op.f('image_id_modelo_fkey')),
+    sa.ForeignKeyConstraint(['id_predio'], ['building.id'], name=op.f('image_id_predio_fkey')),
+    sa.PrimaryKeyConstraint('id', name=op.f('image_pkey'))
+    )
+    op.create_table('building',
+    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
+    sa.Column('id_expedicao', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('complemento', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('descricao', sa.VARCHAR(), autoincrement=False, nullable=True),
+    sa.Column('foto_fachada', sa.VARCHAR(), autoincrement=False, nullable=True),
+    sa.ForeignKeyConstraint(['id_expedicao'], ['expedition.id'], name=op.f('building_id_expedicao_fkey')),
+    sa.PrimaryKeyConstraint('id', name=op.f('building_pkey'))
+    )
+    op.create_table('user',
+    sa.Column('id', sa.INTEGER(), server_default=sa.text("nextval('user_id_seq'::regclass)"), autoincrement=True, nullable=False),
+    sa.Column('email', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('nome_completo', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('senha', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('cargo', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.PrimaryKeyConstraint('id', name='user_pkey'),
+    sa.UniqueConstraint('email', name='user_email_key', postgresql_include=[], postgresql_nulls_not_distinct=False),
+    postgresql_ignore_search_path=False
+    )
+    op.create_table('expedition',
+    sa.Column('id', sa.INTEGER(), autoincrement=True, nullable=False),
+    sa.Column('nome', sa.VARCHAR(length=200), autoincrement=False, nullable=False),
+    sa.Column('localizacao', sa.VARCHAR(), autoincrement=False, nullable=False),
+    sa.Column('data_criacao', sa.DATE(), autoincrement=False, nullable=False),
+    sa.Column('ultima_att', sa.DATE(), autoincrement=False, nullable=True),
+    sa.Column('id_responsavel', sa.INTEGER(), autoincrement=False, nullable=False),
+    sa.Column('descricao', sa.VARCHAR(), autoincrement=False, nullable=True),
+    sa.Column('foto_capa', sa.VARCHAR(), autoincrement=False, nullable=True),
+    sa.ForeignKeyConstraint(['id_responsavel'], ['user.id'], name=op.f('expedition_id_responsavel_fkey')),
+    sa.PrimaryKeyConstraint('id', name=op.f('expedition_pkey'))
+    )
+    # ### end Alembic commands ###
