+++ src/back-end/app/models/expedition.py
@@ -8,8 +8,8 @@ class Expedition(db.Model):
     id = db.Column(db.Integer, primary_key=True, nullable=False)
     nome = db.Column(db.String(200), nullable=False)
     localizacao = db.Column(db.String(), nullable=False)
-    data_criacao = db.Column(db.String(), nullable=False)
-    ultima_att = db.Column(db.String(), nullable=False)
+    data_criacao = db.Column(db.Date, nullable=False)
+    ultima_att = db.Column(db.Date, nullable=True)
     id_responsavel = db.Column(db.Integer, db.ForeignKey('user.id'), nullable=False)
     descricao = db.Column(db.String(), nullable=True)
     foto_capa = db.Column(db.String(), nullable=True)
@@ -20,13 +20,13 @@ class Expedition(db.Model):
     
     # Função para transformar em json
     def as_dict(self):
-        return{
+        return {
             'id': self.id,
             'nome': self.nome,
             'localizacao': self.localizacao,            
-            'data_criacao': self.data_criacao,
+            'data_criacao': self.data_criacao,   
             'ultima_att': self.ultima_att,
             'id_responsavel': self.id_responsavel,
             'descricao': self.descricao,
             'foto_capa': self.foto_capa
-            }
+        }
